Fixed a parameter type error on PointModelFactory

+++ src/Cubes3DIntersection.Application/Factories/PointModelFactory.cs
@@ -4,6 +4,6 @@ namespace Cubes3DIntersection.Application.Factories
 {
     public static class PointModelFactory
     {
-        public static Point3DModel Create(double x0, double y0, int z0) => new Point3DModel(x0, y0, z0);
+        public static Point3DModel Create(double x0, double y0, double z0) => new Point3DModel(x0, y0, z0);
     }
 }
